Rsync files created by create_products.py and add --archive_only_products

+++ src/data/archive.py
@@ -41,38 +41,61 @@ class Archiver:
             self.logger.info("Is cifs://atlas.shore.mbari.org/AUVCTD mounted?")
             sys.exit(1)
         year = self.args.mission.split(".")[0]
-        surveys_dir = os.path.join(surveys_dir, year, "netcdf")
-        self.logger.info(f"Archiving {nc_file_base} files to {surveys_dir}")
-        # To avoid "fchmod failed: Permission denied" message use rsync instead  of cp
-        # https://apple.stackexchange.com/a/206251
-        for ftype in (f"{freq}.nc", "cal.nc", "align.nc", LOG_NAME):
-            src_file = f"{nc_file_base}_{ftype}"
-            dst_file = f"{os.path.join(surveys_dir, os.path.basename(src_file))}"
-            if self.args.clobber and os.path.exists(dst_file):
-                self.logger.info(f"Removing {dst_file}")
-                os.remove(dst_file)
-            if os.path.exists(src_file) and os.path.exists(src_file):
-                if ftype == LOG_NAME:  # log the log file rsync in the log file
-                    self.logger.info(f"rsync {src_file} {surveys_dir}")
-                os.system(f"rsync {src_file} {surveys_dir}")
-                self.logger.info(f"rsync {src_file} {surveys_dir} done.")
-            else:
-                self.logger.debug(f"{src_file} not found")
 
-        # Copy intermediate files to AUVCTD/missionnetcdfs/YYYY/YYYYJJJ
-        YYYYJJJ = "".join(self.args.mission.split(".")[:2])
-        missionnetcdfs_dir = os.path.join(
-            AUVCTD_VOL, MISSIONNETCDFS, year, YYYYJJJ, self.args.mission
-        )
-        Path(missionnetcdfs_dir).mkdir(parents=True, exist_ok=True)
-        src_dir = "/".join(nc_file_base.split("/")[:-1])
-        for log in LOG_FILES:
-            src_file = os.path.join(src_dir, f"{log.replace('.log', '')}.nc")
-            if os.path.exists(src_file):
-                os.system(f"rsync {src_file} {missionnetcdfs_dir}")
-                self.logger.info(f"rsync {src_file} {missionnetcdfs_dir} done.")
+        # To avoid "fchmod failed: Permission denied" message use rsync instead
+        # of cp: https://apple.stackexchange.com/a/206251
+
+        if not self.args.archive_only_products:
+            surveys_dir = os.path.join(surveys_dir, year, "netcdf")
+            self.logger.info(f"Archiving {nc_file_base} files to {surveys_dir}")
+            # Rsync netCDF files to AUVCTD/surveys/YYYY/netcdf
+            for ftype in (f"{freq}.nc", "cal.nc", "align.nc"):
+                src_file = f"{nc_file_base}_{ftype}"
+                dst_file = f"{os.path.join(surveys_dir, os.path.basename(src_file))}"
+                if self.args.clobber and os.path.exists(dst_file):
+                    self.logger.info(f"Removing {dst_file}")
+                    os.remove(dst_file)
+                if os.path.exists(src_file):
+                    os.system(f"rsync {src_file} {surveys_dir}")
+                    self.logger.info(f"rsync {src_file} {surveys_dir} done.")
+                else:
+                    self.logger.debug(f"{src_file} not found")
+
+            # Rsync intermediate files to AUVCTD/missionnetcdfs/YYYY/YYYYJJJ
+            YYYYJJJ = "".join(self.args.mission.split(".")[:2])
+            missionnetcdfs_dir = os.path.join(
+                AUVCTD_VOL, MISSIONNETCDFS, year, YYYYJJJ, self.args.mission
+            )
+            Path(missionnetcdfs_dir).mkdir(parents=True, exist_ok=True)
+            src_dir = "/".join(nc_file_base.split("/")[:-1])
+            for log in LOG_FILES:
+                src_file = os.path.join(src_dir, f"{log.replace('.log', '')}.nc")
+                if os.path.exists(src_file):
+                    os.system(f"rsync {src_file} {missionnetcdfs_dir}")
+                    self.logger.info(f"rsync {src_file} {missionnetcdfs_dir} done.")
+                else:
+                    self.logger.debug(f"{src_file} not found")
+
+        # Rsync files created by create_products.py
+        self.logger.info(f"Archiving product files")
+        for src_dir, dst_dir in (("missionodvs", "odv"), ("missionimages", "images")):
+            src_dir = os.path.join(
+                BASE_PATH, self.args.auv_name, src_dir, self.args.mission
+            )
+            if os.path.exists(src_dir):
+                dst_dir = os.path.join(AUVCTD_VOL, "surveys", year, dst_dir)
+                Path(dst_dir).mkdir(parents=True, exist_ok=True)
+                os.system(f"rsync -r {src_dir}/* {dst_dir}")
+                self.logger.info(f"rsync {src_dir}/* {dst_dir} done.")
             else:
-                self.logger.debug(f"{src_file} not found")
+                self.logger.debug(f"{src_dir} not found")
+
+        # Rsync the processing.log file last so that we get everything
+        src_file = f"{nc_file_base}_{LOG_NAME}"
+        dst_file = f"{os.path.join(surveys_dir, os.path.basename(src_file))}"
+        if os.path.exists(src_file):
+            self.logger.info(f"rsync {src_file} {surveys_dir}")
+            os.system(f"rsync {src_file} {missionnetcdfs_dir}")
 
     def copy_to_M3(self, resampled_nc_file: str) -> None:
         pass
@@ -121,6 +144,11 @@ class Archiver:
             action="store_true",
             help="Remove existing netCDF files before rsyncing to the AUVCTD directory",
         )
+        parser.add_argument(
+            "--archive_only_products",
+            action="store_true",
+            help="Rsync to AUVCTD directory only the products, not the netCDF files",
+        ),
         parser.add_argument(
             "-v",
             "--verbose",
